Simplify decorators and remove unused button from template
Simplified the decorators for child dashboard button and huu match child dashboard button in 'flourish_dashboard_extras.py' file. This change was made to maintain code readability and consistency as there's no need for decorators to be in two lines.
Also, in 'maternal_listboard.html' the unused 'eligibility_button' that was tied to 'result.consent' has been removed to avoid confusion and to maintain cleanliness in the code base.