:recycle: refactor(consent version): Update a consent version to 4 and add associated tests
Updated the default consent version from 3 to 4 in numerous files within the flourish-caregiver and flourish_child modules. This was in accordance with the study plan which proposed a new version of a consent form. Actions include altering imports, function and variable definitions, class properties, and HTML templates. All instances of version '3' have been replaced with '4.'

To accommodate this change, a new test file 'test_subject_reconsent_version4.py' has been added which checks whether participants are able to reconsent under the new version. It mocks various scenarios and checks if participants are successfully enrolled under version 4 of the consent form.

Also, modifications were made to remove and comment on several unrelated fields in 'breastfeeding_questionnaire_admin.py' within the flourish-caregiver module. This appears to be part of a different task related to streamlining the Breastfeeding Questionnaire admin interface.