(Feature): Add user group names for finer dashboard control
This commit introduces the use of user group names into the Flourish Dashboard, specifically within the child_subject and maternal_subject views. By fetching the logged-in user's group names, we extend the context passed to the views. This allows for customizing dashboard buttons and actions based on the user's group. Primary Investigators (PIs) and Coordinators, for instance, get a 'Review' button for the 'IN_PROGRESS' or 'COMPLETE' appointments, while other users cannot access it. Essentially, this is enhancing user role-based permissions within the dashboard.

Signed-off-by: nmunatsibw <mcturnergarai@gmail.com>